+++ rewatch/tests/compile.sh
@@ -34,6 +34,21 @@ else
   exit 1
 fi
 
+# Build from standalone package folder using `rescript build`
+bold "Test: Standalone package can build via rescript from package folder"
+pushd ./packages/standalone > /dev/null
+error_output=$("../../$REWATCH_EXECUTABLE" build 2>&1)
+if [ $? -eq 0 ];
+then
+  success "Standalone package built"
+else
+  error "Error building standalone package"
+  printf "%s\n" "$error_output" >&2
+  popd > /dev/null
+  exit 1
+fi
+popd > /dev/null
+
 node ./packages/main/src/Main.mjs > ./packages/main/src/output.txt
 
 mv ./packages/main/src/Main.res ./packages/main/src/Main2.res