+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Improve option optimization for constants. https://github.com/rescript-lang/rescript/pull/7913
 - Option optimization: do not create redundant local vars. https://github.com/rescript-lang/rescript/pull/7915
 - Js output: remove superfluous newline after every `if`. https://github.com/rescript-lang/rescript/pull/7920
+- Rewatch: Traverse upwards for package resolution in single context projects. https://github.com/rescript-lang/rescript/pull/7896
 
 #### :house: Internal